Develop navDropdown button & fix issues
Create an array which has navbar items details and reduce code to one navItem which can map into the array elements and create dropdown items according to the elements in array. Also there was an error when selecting another item the dropdown, title is changing to default as the page refreshes. As a solution, link Container used. Additionally made the selected item highlighted using active prop